extracted and cleaned up
